rename to directory_reader
