+++ config.w32
@@ -573,7 +573,7 @@ if (PHP_GRPC != "no") {
     "src\\core\\lib\\gprpp\\load_file.cc " +
     "src\\core\\lib\\gprpp\\mpscq.cc " +
     "src\\core\\lib\\gprpp\\per_cpu.cc " +
-    "src\\core\\lib\\gprpp\\posix\\directory.cc " +
+    "src\\core\\lib\\gprpp\\posix\\directory_reader.cc " +
     "src\\core\\lib\\gprpp\\posix\\env.cc " +
     "src\\core\\lib\\gprpp\\posix\\stat.cc " +
     "src\\core\\lib\\gprpp\\posix\\thd.cc " +
@@ -585,7 +585,7 @@ if (PHP_GRPC != "no") {
     "src\\core\\lib\\gprpp\\time_averaged_stats.cc " +
     "src\\core\\lib\\gprpp\\time_util.cc " +
     "src\\core\\lib\\gprpp\\validation_errors.cc " +
-    "src\\core\\lib\\gprpp\\windows\\directory.cc " +
+    "src\\core\\lib\\gprpp\\windows\\directory_reader.cc " +
     "src\\core\\lib\\gprpp\\windows\\env.cc " +
     "src\\core\\lib\\gprpp\\windows\\stat.cc " +
     "src\\core\\lib\\gprpp\\windows\\thd.cc " +
