+++ gRPC-C++.podspec
@@ -835,7 +835,7 @@ Pod::Spec.new do |s|
                       'src/core/lib/gprpp/cpp_impl_of.h',
                       'src/core/lib/gprpp/crash.h',
                       'src/core/lib/gprpp/debug_location.h',
-                      'src/core/lib/gprpp/directory.h',
+                      'src/core/lib/gprpp/directory_reader.h',
                       'src/core/lib/gprpp/dual_ref_counted.h',
                       'src/core/lib/gprpp/env.h',
                       'src/core/lib/gprpp/examine_stack.h',
@@ -1910,7 +1910,7 @@ Pod::Spec.new do |s|
                               'src/core/lib/gprpp/cpp_impl_of.h',
                               'src/core/lib/gprpp/crash.h',
                               'src/core/lib/gprpp/debug_location.h',
-                              'src/core/lib/gprpp/directory.h',
+                              'src/core/lib/gprpp/directory_reader.h',
                               'src/core/lib/gprpp/dual_ref_counted.h',
                               'src/core/lib/gprpp/env.h',
                               'src/core/lib/gprpp/examine_stack.h',
