+++ gRPC-Core.podspec
@@ -1325,7 +1325,7 @@ Pod::Spec.new do |s|
                       'src/core/lib/gprpp/crash.cc',
                       'src/core/lib/gprpp/crash.h',
                       'src/core/lib/gprpp/debug_location.h',
-                      'src/core/lib/gprpp/directory.h',
+                      'src/core/lib/gprpp/directory_reader.h',
                       'src/core/lib/gprpp/dual_ref_counted.h',
                       'src/core/lib/gprpp/env.h',
                       'src/core/lib/gprpp/examine_stack.cc',
@@ -1350,7 +1350,7 @@ Pod::Spec.new do |s|
                       'src/core/lib/gprpp/packed_table.h',
                       'src/core/lib/gprpp/per_cpu.cc',
                       'src/core/lib/gprpp/per_cpu.h',
-                      'src/core/lib/gprpp/posix/directory.cc',
+                      'src/core/lib/gprpp/posix/directory_reader.cc',
                       'src/core/lib/gprpp/posix/env.cc',
                       'src/core/lib/gprpp/posix/stat.cc',
                       'src/core/lib/gprpp/posix/thd.cc',
@@ -1380,7 +1380,7 @@ Pod::Spec.new do |s|
                       'src/core/lib/gprpp/unique_type_name.h',
                       'src/core/lib/gprpp/validation_errors.cc',
                       'src/core/lib/gprpp/validation_errors.h',
-                      'src/core/lib/gprpp/windows/directory.cc',
+                      'src/core/lib/gprpp/windows/directory_reader.cc',
                       'src/core/lib/gprpp/windows/env.cc',
                       'src/core/lib/gprpp/windows/stat.cc',
                       'src/core/lib/gprpp/windows/thd.cc',
@@ -2671,7 +2671,7 @@ Pod::Spec.new do |s|
                               'src/core/lib/gprpp/cpp_impl_of.h',
                               'src/core/lib/gprpp/crash.h',
                               'src/core/lib/gprpp/debug_location.h',
-                              'src/core/lib/gprpp/directory.h',
+                              'src/core/lib/gprpp/directory_reader.h',
                               'src/core/lib/gprpp/dual_ref_counted.h',
                               'src/core/lib/gprpp/env.h',
                               'src/core/lib/gprpp/examine_stack.h',
