+++ grpc.gemspec
@@ -1228,7 +1228,7 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/gprpp/crash.cc )
   s.files += %w( src/core/lib/gprpp/crash.h )
   s.files += %w( src/core/lib/gprpp/debug_location.h )
-  s.files += %w( src/core/lib/gprpp/directory.h )
+  s.files += %w( src/core/lib/gprpp/directory_reader.h )
   s.files += %w( src/core/lib/gprpp/dual_ref_counted.h )
   s.files += %w( src/core/lib/gprpp/env.h )
   s.files += %w( src/core/lib/gprpp/examine_stack.cc )
@@ -1253,7 +1253,7 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/gprpp/packed_table.h )
   s.files += %w( src/core/lib/gprpp/per_cpu.cc )
   s.files += %w( src/core/lib/gprpp/per_cpu.h )
-  s.files += %w( src/core/lib/gprpp/posix/directory.cc )
+  s.files += %w( src/core/lib/gprpp/posix/directory_reader.cc )
   s.files += %w( src/core/lib/gprpp/posix/env.cc )
   s.files += %w( src/core/lib/gprpp/posix/stat.cc )
   s.files += %w( src/core/lib/gprpp/posix/thd.cc )
@@ -1283,7 +1283,7 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/gprpp/unique_type_name.h )
   s.files += %w( src/core/lib/gprpp/validation_errors.cc )
   s.files += %w( src/core/lib/gprpp/validation_errors.h )
-  s.files += %w( src/core/lib/gprpp/windows/directory.cc )
+  s.files += %w( src/core/lib/gprpp/windows/directory_reader.cc )
   s.files += %w( src/core/lib/gprpp/windows/env.cc )
   s.files += %w( src/core/lib/gprpp/windows/stat.cc )
   s.files += %w( src/core/lib/gprpp/windows/thd.cc )
