+++ grpc.gyp
@@ -802,13 +802,13 @@
         'src/core/lib/experiments/experiments.cc',
         'src/core/lib/gprpp/load_file.cc',
         'src/core/lib/gprpp/per_cpu.cc',
-        'src/core/lib/gprpp/posix/directory.cc',
+        'src/core/lib/gprpp/posix/directory_reader.cc',
         'src/core/lib/gprpp/ref_counted_string.cc',
         'src/core/lib/gprpp/status_helper.cc',
         'src/core/lib/gprpp/time.cc',
         'src/core/lib/gprpp/time_averaged_stats.cc',
         'src/core/lib/gprpp/validation_errors.cc',
-        'src/core/lib/gprpp/windows/directory.cc',
+        'src/core/lib/gprpp/windows/directory_reader.cc',
         'src/core/lib/gprpp/work_serializer.cc',
         'src/core/lib/handshaker/proxy_mapper_registry.cc',
         'src/core/lib/http/format_request.cc',
