+++ package.xml
@@ -1210,7 +1210,7 @@
     <file baseinstalldir="/" name="src/core/lib/gprpp/crash.cc" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gprpp/crash.h" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gprpp/debug_location.h" role="src" />
-    <file baseinstalldir="/" name="src/core/lib/gprpp/directory.h" role="src" />
+    <file baseinstalldir="/" name="src/core/lib/gprpp/directory_reader.h" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gprpp/dual_ref_counted.h" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gprpp/env.h" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gprpp/examine_stack.cc" role="src" />
@@ -1235,7 +1235,7 @@
     <file baseinstalldir="/" name="src/core/lib/gprpp/packed_table.h" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gprpp/per_cpu.cc" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gprpp/per_cpu.h" role="src" />
-    <file baseinstalldir="/" name="src/core/lib/gprpp/posix/directory.cc" role="src" />
+    <file baseinstalldir="/" name="src/core/lib/gprpp/posix/directory_reader.cc" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gprpp/posix/env.cc" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gprpp/posix/stat.cc" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gprpp/posix/thd.cc" role="src" />
@@ -1265,7 +1265,7 @@
     <file baseinstalldir="/" name="src/core/lib/gprpp/unique_type_name.h" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gprpp/validation_errors.cc" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gprpp/validation_errors.h" role="src" />
-    <file baseinstalldir="/" name="src/core/lib/gprpp/windows/directory.cc" role="src" />
+    <file baseinstalldir="/" name="src/core/lib/gprpp/windows/directory_reader.cc" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gprpp/windows/env.cc" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gprpp/windows/stat.cc" role="src" />
     <file baseinstalldir="/" name="src/core/lib/gprpp/windows/thd.cc" role="src" />
