+++ src/core/lib/gprpp/windows/directory_reader.cc
@@ -31,7 +31,7 @@
 
 #include <grpc/support/log.h>
 
-#include "src/core/lib/gprpp/directory.h"
+#include "src/core/lib/gprpp/directory_reader.h"
 namespace grpc_core {
 
 namespace {
